Add EmailAlertApi namespace
No code changes in this commit, except for the addition of the
namespace.
